fix: #59 share js files
