+++ src/drawlist.js
@@ -1,0 +1,3 @@
+import './sass/index.scss';
+import './js/drawlist';
+import './js/button_top';
