+++ src/index.js
@@ -1,2 +1,5 @@
 import './sass/index.scss';
-import './js/index';
+import './js/best_slide';
+import './js/button_top';
+import './js/dailylook_move';
+import './js/page_scroll';
