+++ src/js/drawlist.js
@@ -1,0 +1,1 @@
+console.log("hello!")
